Работоспособная версия с выдачей User-токенов

+++ auth/src/main/java/ru/discloud/auth/web/model/UserTokenRemoveRequest.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class UserTokenRemoveRequest {
-    Integer userId;
+    Long userId;
     String deviceId;
 }
